scraping/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
pydantic==2.5.2
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1
